changed 6023 to 6123 table
