+++ src/test/scala/com/hhi/sap/analysis/ANL_THD_MRPL_WEEKTest.scala
@@ -13,7 +13,7 @@ class ANL_THD_MRPL_WEEKTest extends FlatSpec with SparkSessionTestWrapper{
   private val INPUTPATH = "src/test/resources"
   private val TABLE4 = "/output/table4"
   private val FILENPATH_ZPSCT600 = "/ZPDCV6021/*"
-  private val FILENPATH_ZPDCT6023 = "/ZPDCT6023/*"
+  private val FILENPATH_ZPDCT6123 = "/ZPDCT6123/*"
   private val FILEPATH_EBAN = "/EBAN/*"
   private val FILEPATH_MARA = "/MARA/*"
 
@@ -38,8 +38,6 @@ class ANL_THD_MRPL_WEEKTest extends FlatSpec with SparkSessionTestWrapper{
   }
 
   "simple code3" should "be tested." in {
-    import ss.sqlContext.sparkSession.implicits._
-
     ss.sparkContext.parallelize(List(
       ("West",  "Apple1",  -9.0, 10),
       ("West",  "Orange", -8.0, 15),
@@ -68,21 +66,18 @@ class ANL_THD_MRPL_WEEKTest extends FlatSpec with SparkSessionTestWrapper{
       .map{ case ((store, prod), (amt, units)) => (store, prod, amt, units) }
       .groupBy(-3 until 10 contains _._3)
       .foreach(println)
-//      .toDF("store", "prod", "amt", "units")
-//      .where("amt < -1 ")
-//      .show()
   }
 
-  "ZPDCT6023" should "be counted." in new SparkFileReader {
-    println(getFolder(INPUTPATH+FILENPATH_ZPDCT6023).count())
+  "ZPDCT6123" should "be counted." in new SparkFileReader {
+    println(getFolder(INPUTPATH+FILENPATH_ZPDCT6123).count())
   }
 
-  "ZPDCT6023 " should "make dataframe." in new SparkFileReader {
-    //"Please be generated a table ZPDCT6023 on class ANL_THD_ZPSCT600_RTest "
-    val tb_ZPDCT6023 = ss.read.option("header", "true").csv(INPUTPATH + TABLE4)
+  "ZPDCT6123 " should "make dataframe." in new SparkFileReader {
+    //"Please be generated a table ZPDCT6123 on class ANL_THD_ZPSCT600_RTest "
+    val tb_ZPDCT6123 = ss.read.option("header", "true").csv(INPUTPATH + TABLE4)
 
-    new ANL_THD_MRPL_WEEK(ss.sqlContext).run(tb_ZPDCT6023
-      .select(TERM_MASTER.ZPDCT6023.COMPANYID, TERM_MASTER.ZPDCT6023.SAUPBU, TERM_MASTER.ZPDCT6023.ZTRKNO, TERM_MASTER.ZPDCT6023.ZREVNO, TERM_MASTER.ZPDCT6023.PSPID, TERM_MASTER.ZPDCT6023.STG_GUBUN, TERM_MASTER.ZPDCT6023.MAT_GUBUN, TERM_MASTER.ZPDCT6023.WEEK))
+    new ANL_THD_MRPL_WEEK(ss.sqlContext).run(tb_ZPDCT6123
+      .select(TERM_MASTER.ZPDCT6123.COMPANYID, TERM_MASTER.ZPDCT6123.SAUPBU, TERM_MASTER.ZPDCT6123.PSPID, TERM_MASTER.ZPDCT6123.STG_GUBUN, TERM_MASTER.ZPDCT6123.MAT_GUBUN, TERM_MASTER.ZPDCT6123.WEEK))
       .show()
   }
 }
